automatically backed up by learn

+++ db/migrate/01_create_table.rb
@@ -5,7 +5,7 @@ class CreateTable < ActiveRecord::Migration[4.2]
    create_table :users do |u|
      u.string :username
      u.string :password
-     u.integer :balance
+     u.decimal :balance
    end
   end
